분배로직 수정

+++ src/main/java/com/plateer/ec1/order/vo/OrderBenefitVO.java
@@ -18,6 +18,6 @@ public class OrderBenefitVO extends OrderBenefitBaseVO {
     List<OrderProductBaseVO> orderProductBaseVOList;
 
     public long distributeAplyAmt(long prdBnfAplyOrdPrc, long totalPrdBnfApyOrdPrc){
-        return (prdBnfAplyOrdPrc / totalPrdBnfApyOrdPrc) * this.getAplyAmt();
+        return (long) (((double) prdBnfAplyOrdPrc / (double) totalPrdBnfApyOrdPrc) * this.getAplyAmt());
     }
 }
